Rudimentary Drawing added to the game
Not all constraints are in place, but you can click on stars and make lines. Changed the nodestar, ritual, and energy to use start instead of awake to prevent null pointers. Added the star script to the star class script. Made a linerenderer and added functionality to instantiate into the scene at the proper time